virNetDevVethCreate: assign names based on mac address by default
Interface names do not have to be numerical (or veth + number) and trying to
assign them to that format is susceptible to race conditions.  Instead,
assign the parent interface name according to the mac address (the last
three bytes) if no name was given by the caller and use the parent interface
name plus 'p' for the container name.

Signed-off-by: Oskari Saarenmaa <os@ohmu.fi>