virNetDevVethCreate: assign container if name based on parent if name
Replace the loop trying to find a free veth interface name for the container
by assigning the container if name to parent name + 'p' by default.
Interface name selection logic is susceptible to race conditions, so try to
select just one name by default and use that as a template for the second
name.  The parent name can also be overriden in domain configuration.

Signed-off-by: Oskari Saarenmaa <os@ohmu.fi>